+++ docs/changelog.md
@@ -1,3 +1,8 @@
+### 1.4.1 - Improved and documented validation feature 
+
+ * A validator function may now return `None`
+ * improved documentation on `@validate` in particular for custom validators implementation
+
 ### 1.4.0 - Added validation feature
 
  * New: `@validate` annotation allowing to implement input validation. Comes with a bunch of built-in validators.